Initial refactoring of country selection code logic and other remaining custom hooks
