+++ src/constants/dataAndFeatures.js
@@ -2,6 +2,34 @@ import screenshotExpenses from '@/images/screenshots/expenses.png'
 import screenshotPayroll from '@/images/screenshots/payroll.png'
 import screenshotReporting from '@/images/screenshots/reporting.png'
 import screenshotVatReturns from '@/images/screenshots/vat-returns.png'
+import { ArgentinaFlag, ColombiaFlag, MexicoFlag, SpainFlag, ChileFlag } from '@/components/Flags'
+
+export const LIST_OF_CONTRIES = [
+  'España', 'Colombia', 'Mexico', 'Argentina', 'Chile'
+]
+
+export const SELECTED_COUNTRIES_STATS = {
+  España: {
+    name: 'España',
+    flag: <SpainFlag />
+  },
+  Colombia: {
+    name: 'Colombia',
+    flag: <ColombiaFlag />
+  },
+  Mexico: {
+    name: 'Mexico',
+    flag: <MexicoFlag />
+  },
+  Argentina: {
+    name: 'Argentina',
+    flag: <ArgentinaFlag />
+  },
+  Chile: {
+    name: 'Chile',
+    flag: <ChileFlag />
+  }
+}
 
 export const features = [
   {
